refactor: file 구조, 이름 변경
