+++ client/src/components/login/index.tsx
@@ -1,0 +1,11 @@
+import LoginForm from '@/components/login/Form';
+import LoginLayout from '@/components/login/Layout';
+
+function Login() {
+  return (
+    <LoginLayout>
+      <LoginForm />
+    </LoginLayout>
+  );
+}
+export default Login;
